cleaning up and adding metadata
